coderabbit

+++ samigo/samigo-services/src/java/org/sakaiproject/tool/assessment/facade/AssessmentFacadeQueries.java
@@ -849,7 +849,7 @@ public class AssessmentFacadeQueries extends HibernateDaoSupport implements Asse
 	public List getQuestionsIdList(final Long assessmentId) {
 		HibernateCallback<List<Long>> hcb = session -> session
 				.createQuery("select i.itemId from ItemData i, SectionData s,  AssessmentData a where a = s.assessment and s = i.section and a.assessmentBaseId = :id")
-				.setParameter("id", assessmentId.longValue())
+				.setParameter("id", assessmentId)
 				.list();
 		return getHibernateTemplate().execute(hcb);
 	}
